Improve autocorrelation plumbing to HFMetrics
Also tailor HFMetricsFilter to Host implementation using LaneArray (including
vector operators).

+++ src/basecaller/dataTypes/HQRFPhysicalStates.h
@@ -32,7 +32,7 @@ namespace PacBio {
 namespace Mongo {
 namespace Data {
 
-enum class HQRFPhysicalStates : std::uint8_t
+enum class HQRFPhysicalStates : std::uint16_t
 {
     EMPTY = 0,
     SINGLE,
